Update WMIDnsServer.ps1
Removed some outputs that from $Hashtable.Add() and other extraneous outputs.  
Modified the switch statements in Get-WmiDNSZone after finding the accurate data in technet for the values of certain fields.
Added "Name" to the default display set in Get-WmiDNSResourceRecordSet, which is the full FQDN, which is provided by most other DNS modules, even thought it is not included in the output for Microsoft DNS - it will make scripting against this easier.